% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{credentials}
\alias{credentials}
\title{Detect mathpix credentials}
\usage{
credentials()
}
\value{
a list of detected credentials (or this package's credentials)
}
\description{
Checks environmental variables for \code{MATHPIX_APP_ID} and \code{MATHPIX_APP_KEY} values.
}
