% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{mathpix_api}
\alias{mathpix_api}
\title{Connect to the mathpix API and translate an image to LaTeX}
\usage{
mathpix_api(img)
}
\arguments{
\item{img}{image to be converted to LaTeX}
}
\value{
a character string of LaTeX commands (or NULL if fails).
}
\description{
Connect to the mathpix API and translate an image to LaTeX
}
\details{
You must save your own API key in your
  environment (e.g. `~/.Renviron`) with the identifiers `MATHPIX_APP_ID` and
  `MATHPIX_APP_KEY`. This can be tested with `mathpix:::credentials()`.
}
\keyword{internal}
