% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{render_latex}
\alias{render_latex}
\title{Convert a LaTeX block to a PNG image}
\usage{
render_latex(latex, file = NULL)
}
\arguments{
\item{latex}{LaTeX code to be evaluated}

\item{file}{filename (with directory) to save the image to (defaults to `/tmp/tempfile()`)}
}
\value{
(invisibly) the filename (with directory) where the image was saved.
}
\description{
Convert a LaTeX block to a PNG image
}
