% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{rmarkdown_block}
\alias{rmarkdown_block}
\title{Add LaTeX commands as an \code{rmarkdown} equation}
\usage{
rmarkdown_block(img, retry = FALSE)
}
\arguments{
\item{img}{image to be converted to LaTeX}

\item{retry}{If Mathpix is not able to process the image, should we try again with
a re-processed image?}
}
\value{
an \code{rmarkdown} equation block
}
\description{
Converts an image to LaTeX then inserts that into an \code{rmarkdown} equation
(surrounded by $$).
}
\details{
You must save your own API key in your
environment (e.g. \verb{~/.Renviron}) with the identifiers \code{MATHPIX_APP_ID} and
\code{MATHPIX_APP_KEY}. This can be tested with \code{mathpix:::credentials()}.
}
\keyword{internal}
