% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{render_latex}
\alias{render_latex}
\title{Convert a LaTeX block to a PNG image}
\usage{
render_latex(latex, fileDir = NULL, ...)
}
\arguments{
\item{latex}{LaTeX code to be evaluated}

\item{fileDir}{directory in which to save the image to (defaults to `/tmp/tempfile()`)}

\item{...}{other options to pass to \code{\link[texPreview]{texPreview}}.}
}
\value{
NULL (invisibly)
}
\description{
Convert a LaTeX block to a PNG image
}
