% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{render_latex}
\alias{render_latex}
\title{Convert a 'LaTeX' expression to an image (render)}
\usage{
render_latex(latex, fileDir = NULL, ...)
}
\arguments{
\item{latex}{'LaTeX' code to be evaluated. Surround in \code{$} or \code{$$}.}

\item{fileDir}{directory in which to save the image to (defaults to
`/tmp/tempfile()`).}

\item{...}{other options to pass to \code{tex_preview}.}
}
\value{
NULL (invisibly)
}
\description{
This calls \code{tex_preview} to render a 'LaTeX'
expression into an image, either as a temporary file or saved to disk.
}
