% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{rmarkdown_block}
\alias{rmarkdown_block}
\title{Add LaTeX commands as an \code{rmarkdown} equation}
\usage{
rmarkdown_block(img, trial = FALSE)
}
\arguments{
\item{img}{image to be converted to LaTeX}

\item{trial}{should the trial API key be used (see Details)?}
}
\value{
an \code{rmarkdown} equation block
}
\description{
Converts an image to LaTeX then inserts that into an \code{rmarkdown} equation
(surrounded by $$).
}
\details{
I have obtained an API key for use with this app, which I have
included. If you have your own API key feel free to save that in your
environment (e.g. \code{~/.Renviron}) with the identifiers \code{MATHPIX_APP_ID} and
\code{MATHPIX_APP_KEY}. If this fails for some reason, the \code{trial} API key can
be used (as found on the mathpix API documentation site).
}
\keyword{internal}
